callbacks: give up on named callbacks (for the time being)
Until overriding heterogeneous variadic methods is allowed the
implementation of named callbacks is unviable.  Attempts to circumvent
the compilers wishes in this regard lead to linker errors.

I can't see why the restriction applies as the description of how they
work (in the book) don't seem to make it impossible.

Signed-off-by: Peter Williams <pwil3058@gmail.com>